requirements.txt
python-telegram-bot>=20.0
APScheduler>=3.10
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
python-docx>=1.1.0